Review changes done

+++ assignment2/src/main/java/org/example/CompoundInterest.java
@@ -2,11 +2,11 @@ package org.example;
 
 public class CompoundInterest extends Interest {
 
-    public CompoundInterest(float principal, float rate, int period) {
+    public CompoundInterest(float principal, float rate, float period) {
         super(principal, rate, period);
     }
 
-    public double getFinalAmount() {
-        return Principal * Math.pow(1 + Rate, Period);
+    public float getFinalAmount() {
+        return (float) ( Principal * Math.pow(1 + Rate, Period));
     }
 }
