+++ assignment2/src/main/java/org/example/SimpleInterest.java
@@ -2,12 +2,11 @@ package org.example;
 
 public class SimpleInterest extends Interest {
 
-    public SimpleInterest(float principal, float rate, int period) {
+    public SimpleInterest(float principal, float rate, float period) {
         super(principal, rate, period);
     }
 
-
-    public double getFinalAmount() {
-        return Principal + (Principal * Rate * Period)/100;
+    public float getFinalAmount() {
+        return (float) (Principal + (Principal * Rate * Period)/100);
     }
 }
